refactor: restore AuthProvider component with email whitelist check and user role fetching
